Day 06 - Star 2

+++ Day06.Tests/TestWinCalculator.cs
@@ -10,7 +10,7 @@ namespace Day06.Tests
         [TestCase(7, 9, ExpectedResult = 4)]
         [TestCase(15, 40, ExpectedResult = 8)]
         [TestCase(30, 200, ExpectedResult = 9)]
-        public int Test(int time, int distance)
+        public long Test(long time, long distance)
         {
             var x = new WinCalculator();
             return x.WaysToWin(time, distance);
