+++ Day06/WinCalculator.cs
@@ -2,7 +2,7 @@
 {
     public class WinCalculator
     {
-        public int WaysToWin(int time, int distance)
+        public long WaysToWin(long time, long distance)
         {
             // Solve: x^2 - time * x + distance > 0
             var discriminant = time * time - 4 * (distance + 1);
@@ -10,7 +10,7 @@
             var high = (time + rootD) / 2;
             var low = (time - rootD) / 2;
 
-            var waysToWin = (int)(Math.Floor(high) - Math.Ceiling(low)) + 1;
+            var waysToWin = (long)(Math.Floor(high) - Math.Ceiling(low)) + 1;
 
             return waysToWin;
         }
